+++ pkg/alertmanager/types.go
@@ -368,6 +368,9 @@ type discordConfig struct {
 	WebhookURL    string            `yaml:"webhook_url,omitempty"`
 	Title         string            `yaml:"title,omitempty"`
 	Message       string            `yaml:"message,omitempty"`
+	Content       string            `yaml:"content,omitempty"`
+	Username      string            `yaml:"username,omitempty"`
+	AvatarURL     string            `yaml:"avatar_url,omitempty"`
 }
 
 type webexConfig struct {