+++ CHANGELOG.md
@@ -5,7 +5,8 @@
 * [FEATURE] Add `keepIdentifyingResourceAttributes` field in OTLPConfig for `Prometheus` and `PrometheusAgent` CRDs. #7240
 * [FEATURE] Add `fallbackScrapeProtocol` field to ScrapeClass for `Prometheus` and `PrometheusAgent` CRDs. #7288
 * [FEATURE] Support `MSTeamsV2Config` receiver in Alertmanager configuration Secret. #7296
-* [FEATURE] Add `Timeout` field in Webhook receiver in Alertmanager configuration Secret. #7312
+* [FEATURE] Add `timeout` field in Webhook receiver in Alertmanager configuration Secret. #7312
+* [FEATURE] Add `content`, `username` and `avatarUrl` fields in Discord receiver in Alertmanager configuration secret. #7313
 * [FEATURE] Add `scrapeFailureLogFile` field to the Prometheus and PrometheusAgent CRDs. #7038 #7311
 * [BUGFIX] Validate respective api URL field in PagerDutyConfigs, DiscordConfigs, SlackConfigs and WebhookConfigs in AlertManagerConfig CRD. #6585